ref: split opencode build and publish script
