+++ packages/opencode/script/build.ts
@@ -1,0 +1,87 @@
+#!/usr/bin/env bun
+const dir = new URL("..", import.meta.url).pathname
+process.chdir(dir)
+import { $ } from "bun"
+import pkg from "../package.json"
+
+const version = process.env["OPENCODE_VERSION"]!
+
+console.log(`building ${version}`)
+
+const GOARCH: Record<string, string> = {
+  arm64: "arm64",
+  x64: "amd64",
+  "x64-baseline": "amd64",
+}
+
+const targets = [
+  ["windows", "x64"],
+  ["linux", "arm64"],
+  ["linux", "x64"],
+  ["linux", "x64-baseline"],
+  ["darwin", "x64"],
+  ["darwin", "x64-baseline"],
+  ["darwin", "arm64"],
+]
+
+await $`rm -rf dist`
+
+const optionalDependencies: Record<string, string> = {}
+for (const [os, arch] of targets) {
+  console.log(`building ${os}-${arch}`)
+  const name = `${pkg.name}-${os}-${arch}`
+  await $`mkdir -p dist/${name}/bin`
+  await $`CGO_ENABLED=0 GOOS=${os} GOARCH=${GOARCH[arch]} go build -ldflags="-s -w -X main.Version=${version}" -o ../opencode/dist/${name}/bin/tui ../tui/cmd/opencode/main.go`.cwd(
+    "../tui",
+  )
+  await $`bun build --define OPENCODE_TUI_PATH="'../../../dist/${name}/bin/tui'" --define OPENCODE_VERSION="'${version}'" --compile --target=bun-${os}-${arch} --outfile=dist/${name}/bin/opencode ./src/index.ts`
+
+  // smoke test only on matching OS/arch
+  if (
+    process.platform === (os === "windows" ? "win32" : os) &&
+    (process.arch === arch || (process.arch === "x64" && arch === "x64-baseline"))
+  ) {
+    console.log(`smoke test: running dist/${name}/bin/opencode --version`)
+    await $`./dist/${name}/bin/opencode --version`
+  }
+
+  await $`rm -rf ./dist/${name}/bin/tui`
+  await Bun.file(`dist/${name}/package.json`).write(
+    JSON.stringify(
+      {
+        name,
+        version,
+        os: [os === "windows" ? "win32" : os],
+        cpu: [arch],
+      },
+      null,
+      2,
+    ),
+  )
+  optionalDependencies[name] = version
+}
+
+// write meta-package
+await $`mkdir -p ./dist/${pkg.name}`
+await $`cp -r ./bin ./dist/${pkg.name}/bin`
+await $`cp ./script/postinstall.mjs ./dist/${pkg.name}/postinstall.mjs`
+await Bun.file(`./dist/${pkg.name}/package.json`).write(
+  JSON.stringify(
+    {
+      name: pkg.name + "-ai",
+      bin: {
+        [pkg.name]: `./bin/${pkg.name}`,
+      },
+      scripts: {
+        postinstall: "node ./postinstall.mjs",
+      },
+      version,
+      optionalDependencies,
+    },
+    null,
+    2,
+  ),
+)
+
+// save optionalDependencies separately for publish step
+await Bun.file("./dist/optionalDependencies.json").write(JSON.stringify(optionalDependencies, null, 2))
